refactor: Move actions into separate package
