+++ packages/cohort/src/bank_api.rs
@@ -1,13 +1,10 @@
-use async_trait::async_trait;
-use std::fmt;
-use std::fmt::{Display, Formatter};
 use std::sync::Arc;
 
-use deadpool_postgres::GenericClient;
 use rusty_money::iso::Currency;
 use rusty_money::Money;
-use tokio_postgres::types::ToSql;
 
+use crate::actions::account_update::AccountUpdate;
+use crate::actions::transfer::Transfer;
 use crate::model::bank_account::BankAccount;
 use crate::state::model::AccountUpdateRequest;
 use crate::state::postgres::data_store::DataStore;
@@ -55,108 +52,5 @@ impl BankApi {
 
     pub async fn transfer_one(db: Arc<Database>, action: Transfer) -> Result<u64, String> {
         action.execute(&db.pool.get().await.unwrap()).await
-    }
-}
-
-#[derive(Debug)]
-pub struct AccountUpdate {
-    pub data: AccountUpdateRequest,
-    pub action: &'static str,
-    pub new_version: u64,
-}
-
-impl Display for AccountUpdate {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "AccountUpdate: [action: {}, account: {}, amount: {}]",
-            self.action, self.data.account, self.data.amount,
-        )
-    }
-}
-
-impl AccountUpdate {
-    pub fn deposit(data: AccountUpdateRequest, new_version: u64) -> Self {
-        Self {
-            data: AccountUpdateRequest::new(data.account, data.amount.replace(['-', '+'], "")),
-            action: "Deposit",
-            new_version,
-        }
-    }
-
-    pub fn withdraw(data: AccountUpdateRequest, new_version: u64) -> Self {
-        Self {
-            data: AccountUpdateRequest::new(data.account, format!("-{}", data.amount.replace(['-', '+'], ""))),
-            action: "Withdraw",
-            new_version,
-        }
-    }
-
-    fn sql() -> &'static str {
-        r#"
-            UPDATE bank_accounts SET data = data ||
-            jsonb_build_object(
-                'amount', ((data->>'amount')::DECIMAL + (($1)::TEXT)::DECIMAL)::TEXT,
-                'talosState', jsonb_build_object('version', ($2)::BIGINT)
-            )
-            WHERE "number" = $3 AND (data->'talosState'->'version')::BIGINT <= ($2)::BIGINT
-        "#
-    }
-}
-
-#[async_trait]
-impl Action for AccountUpdate {
-    async fn execute<T>(&self, client: &T) -> Result<u64, String>
-    where
-        T: GenericClient + Sync,
-    {
-        let params: &[&(dyn ToSql + Sync)] = &[&self.data.amount, &(self.new_version as i64), &self.data.account];
-
-        let statement = client.prepare_cached(Self::sql()).await.unwrap();
-        client.execute(&statement, params).await.map_err(|e| e.to_string())
-    }
-}
-
-#[derive(Debug)]
-pub struct Transfer {
-    pub from: String,
-    pub to: String,
-    pub amount: String,
-    pub new_version: u64,
-}
-
-impl Display for Transfer {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(f, "Transfer: [from: {}, to: {}, amount: {}]", self.from, self.to, self.amount)
-    }
-}
-
-impl Transfer {
-    pub fn new(from: String, to: String, amount: String, new_version: u64) -> Transfer {
-        Transfer { from, to, amount, new_version }
-    }
-
-    fn sql() -> &'static str {
-        r#"
-        UPDATE bank_accounts SET data = COALESCE(
-            CASE
-                WHEN "number" = ($2)::TEXT AND (data->'talosState'->>'version')::BIGINT <= ($3)::BIGINT THEN data || jsonb_build_object('amount', ((data->>'amount')::DECIMAL - (($1)::TEXT)::DECIMAL)::TEXT, 'talosState', jsonb_build_object('version', ($3)::BIGINT))
-                WHEN "number" = ($4)::TEXT AND (data->'talosState'->>'version')::BIGINT <= ($5)::BIGINT THEN data || jsonb_build_object('amount', ((data->>'amount')::DECIMAL + (($1)::TEXT)::DECIMAL)::TEXT, 'talosState', jsonb_build_object('version', ($5)::BIGINT))
-            END, data)
-        WHERE "number" in(($2)::TEXT, ($4)::TEXT)
-        "#
-    }
-}
-
-#[async_trait]
-impl Action for Transfer {
-    async fn execute<T>(&self, client: &T) -> Result<u64, String>
-    where
-        T: GenericClient + Sync,
-    {
-        let params: &[&(dyn ToSql + Sync)] = &[&self.amount, &self.from, &(self.new_version as i64), &self.to, &(self.new_version as i64)];
-
-        let statement = client.prepare_cached(Self::sql()).await.unwrap();
-        client.execute(&statement, params).await.map_err(|e| e.to_string())
     }
 }
