+++ packages/cohort/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod actions; // TODO: move to /packages/src/core/
 pub mod bank_api;
 pub mod config_loader;
 pub mod core; // TODO: move to /packages/src/core/
